+++ src/main/java/org/iq47/model/ItemRepository.java
@@ -9,7 +9,7 @@ import java.util.Collection;
 public interface ItemRepository extends JpaRepository<Item, Long> {
     Collection<Item> getItemsByName(String name);
     Item getItemByName(String name);
-    Item getItemById(long id);
+    Item getItemById(int id);
     Collection<Item> getItemsByNameStartsWithIgnoreCase(String name);
     Collection<Item> getTop5ItemsByNameContainsIgnoreCase(String name);
 }
